requirements.txt
asgiref
black
certifi
cffi
charset-normalizer
click
coreapi
coreschema
cryptography
defusedxml
Django
django-cors-headers
django-filter
django-phonenumber-field
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
drf-nested-routers
drf-yasg
idna
inflection
isort
itypes
Jinja2
MarkupSafe
mypy-extensions
oauthlib
packaging
pathspec
phonenumbers
Pillow
platformdirs
psycopg2-binary
pycparser
PyJWT
pyparsing
python-dotenv
python3-openid
pytz
requests
requests-oauthlib
ruamel.yaml
ruamel.yaml.clib
six
social-auth-app-django
social-auth-core
sqlparse
tomli
uritemplate
urllib3
